+++ server/routes/orgCrime.js
@@ -15,8 +15,8 @@ const findAndCatgeorizeOrgCrimes = async () => {
     .populate('ownerAlliance', 'name')
     .populate('roles.owner', 'name');
 
-  const orgCrimes = allOrgCrimes.filter((crime) => !crime.owner);
-  const claimedOwnOrgCrimes = allOrgCrimes.filter((crime) => !!crime.owner);
+  const orgCrimes = allOrgCrimes.filter((crime) => !crime.ownerAlliance);
+  const claimedOwnOrgCrimes = allOrgCrimes.filter((crime) => crime.ownerAlliance);
 
   return { orgCrimes, claimedOwnOrgCrimes };
 };
@@ -40,6 +40,26 @@ router.get('/', async (req, res) => {
   });
 });
 
+const claimOrgCrimeCriterias = (user, orgCrime, batteryCost, now) => {
+  if (!user || !orgCrime) {
+    return 'Something went wrong';
+  }
+  if (orgCrime.gracePeriod > now) {
+    return 'This crime is currently unavailable';
+  }
+  if (!user.alliance) {
+    return 'You need to be in an alliance to do organized crimes';
+  }
+  if (orgCrime.owner) {
+    return 'This crime is already claimed by someone';
+  }
+  if (user.playerStats.battery < batteryCost) {
+    return 'Insufficent battery';
+  }
+
+  return null;
+};
+
 // @PUT
 // PRIVATE
 // Claims the organized crime
@@ -47,11 +67,24 @@ router.put('/', async (req, res) => {
   const now = Date.now();
   const userId = req.user._id;
   const { crimeId } = req.body;
-  const user = await User.findById(userId).select('alliance').lean();
+  const user = await User.findById(userId);
   const orgCrime = await OrgCrime.findById(crimeId);
+  const batteryCost = 10;
+
+  const disallowed = claimOrgCrimeCriterias(user, orgCrime, batteryCost, now);
+
+  if (disallowed) {
+    return res.status(403).json({
+      success: false,
+      message: disallowed,
+    });
+  }
+
   orgCrime.claimOwner(user._id, user.alliance, now);
   await orgCrime.save();
 
+  user.batteryDrain(batteryCost);
+  const updatedUser = await saveAndUpdateUser(user);
   const allOrgCrimes = await findAndCatgeorizeOrgCrimes();
 
   return res.status(200).json({
@@ -59,8 +92,31 @@ router.put('/', async (req, res) => {
     message: 'Organized Crime claimed..',
     orgCrimes: allOrgCrimes.orgCrimes,
     claimedOwnOrgCrimes: allOrgCrimes.claimedOwnOrgCrimes,
+    user: updatedUser,
   });
 });
+
+const claimOrgCrimeRoleCriterias = (user, orgCrime, batteryCost, now, roleName) => {
+  if (!user || !orgCrime) {
+    return 'Something went wrong';
+  }
+  if (orgCrime.gracePeriod > now) {
+    return 'This crime is currently unavailable';
+  }
+  if (!user.alliance) {
+    return 'You need to be in an alliance to do organized crimes';
+  }
+  if (user.alliance.toString() !== orgCrime.ownerAlliance.toString()) {
+    return 'This crime is claimed by another alliance';
+  }
+  const role = orgCrime.roles.find((r) => r.roleName === roleName);
+  if (role.owner) {
+    return 'This role is already claimed';
+  }
+  if (user.playerStats.batteryCost < batteryCost) {
+    return 'Insufficent battery';
+  }
+};
 
 // @PATCH
 // PRIVATE
@@ -68,11 +124,26 @@ router.put('/', async (req, res) => {
 router.patch('/', async (req, res) => {
   const userId = req.user._id;
   const { crimeId, roleName } = req.body;
-  const user = await User.findById(userId).select('alliance').lean();
+  const user = await User.findById(userId);
   const orgCrime = await OrgCrime.findById(crimeId);
+  const now = Date.now();
+
+  const userAlreadyInCrime = orgCrime.roles.some((role) => role.owner.toString() === userId.toString());
+
+  const batteryCost = userAlreadyInCrime ? 0 : 5;
+  const disallowed = claimOrgCrimeRoleCriterias(user, orgCrime, batteryCost, now, roleName);
+
+  if (disallowed) {
+    return res.status(403).json({
+      success: false,
+      message: disallowed,
+    });
+  }
+
   orgCrime.claimRole(user._id, roleName);
   await orgCrime.save();
 
+  const updatedUser = await saveAndUpdateUser(user);
   const allOrgCrimes = await findAndCatgeorizeOrgCrimes();
 
   return res.status(200).json({
@@ -80,6 +151,7 @@ router.patch('/', async (req, res) => {
     message: `${orgCrime.name} role ${roleName} claimed..`,
     orgCrimes: allOrgCrimes.orgCrimes,
     claimedOwnOrgCrimes: allOrgCrimes.claimedOwnOrgCrimes,
+    user: updatedUser,
   });
 });